requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
python-decouple==3.6
pytz==2022.5
sqlparse==0.4.3
tzdata==2022.5
pytest==7.1.3
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-forked==1.4.0